renamed website blueprint to core blueprint
