+++ app/core/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("core", __name__)
+
+from app.core import routes
