--- app/website/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint("website", __name__)
-
-from app.website import routes
